crcN fully functional, removed unneeded input and cleaned up comments
In addition the max crc length for a two-byte integer was set to 15.
This is because a 16-bit crc has a polynomial of length of 17 which is
larger than the data it is operating on.  A 16-bit crc may also be
excessive for a 16-bit data type.  No real impact to the code's
usefulness is expected by this change