Changed function call back to including 3 inputs and set up code for debugging
Code is not totally operational yet, so it was set up for debugging to
be done later.  Also I included a dataSize variable in the function call
because having the library create a vector object to determine the size
of the data input was proving to not be beneficial enough for the
effort.  It is just as effective to force the use to utilize a
sizeof(data)/2 statement and the resulting code will be easier to
implement and take up less memory on the ATMega328.